requirements.txt
setuptools
numpy
neat-python
graphviz
matplotlib
gym
flask
pytest 
pandas
coverage
pylint
httpx
# elasticbeanstalk
fastapi 
gunicorn
uvicorn

